requirements.txt
torch>=1.9.0
timm>=0.4.12
numpy>=1.19.2
huggingface_hub>=0.11.0
requests>=2.26.0
Pillow>=8.0.0
ftfy>=6.0.0
regex>=2022.3.15 